fix: 안쓰는 코드 정리
